+++ Cargo.toml
@@ -17,6 +17,7 @@ futures = "0.3"
 gas-estimation = { git = "https://github.com/cowprotocol/gas-estimation", tag = "v0.7.2", features = ["web3_", "tokio_"] }
 hex = { version = "0.4", default-features = false }
 hex-literal = "0.4"
+hyper = "0.14"
 itertools = "0.11"
 lazy_static = "1"
 maplit = "1"
