+++ crates/shared/src/trade_finding/external.rs
@@ -45,7 +45,7 @@ impl ExternalTradeFinder {
             amount: query.in_amount,
             kind: query.kind,
             deadline,
-            log_id: crate::tracing::get_task_local_storage(),
+            log_id: crate::request_id::get_task_local_storage(),
         };
 
         let body = serde_json::to_string(&order).context("failed to encode body")?;
